backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-pptx==0.6.23
openai>=1.12.0
python-dotenv==1.0.0
lxml==4.9.3
requests==2.31.0
Pillow==10.1.0
